Move 3D scene import classes to their own folder
