+++ editor/import/3d/editor_import_collada.h
@@ -31,7 +31,7 @@
 #ifndef EDITOR_IMPORT_COLLADA_H
 #define EDITOR_IMPORT_COLLADA_H
 
-#include "editor/import/resource_importer_scene.h"
+#include "editor/import/3d/resource_importer_scene.h"
 
 class EditorSceneFormatImporterCollada : public EditorSceneFormatImporter {
 	GDCLASS(EditorSceneFormatImporterCollada, EditorSceneFormatImporter);
