+++ editor/import/3d/post_import_plugin_skeleton_renamer.h
