+++ editor/import/3d/post_import_plugin_skeleton_track_organizer.cpp
@@ -30,7 +30,7 @@
 
 #include "post_import_plugin_skeleton_track_organizer.h"
 
-#include "editor/import/scene_import_settings.h"
+#include "editor/import/3d/scene_import_settings.h"
 #include "scene/3d/skeleton_3d.h"
 #include "scene/animation/animation_player.h"
 #include "scene/resources/bone_map.h"
