+++ editor/import/SCsub
@@ -3,3 +3,4 @@
 Import("env")
 
 env.add_source_files(env.editor_sources, "*.cpp")
+env.add_source_files(env.editor_sources, "3d/*.cpp")
