+++ modules/gltf/editor/editor_scene_importer_blend.h
@@ -34,7 +34,7 @@
 #ifdef TOOLS_ENABLED
 
 #include "editor/editor_file_system.h"
-#include "editor/import/resource_importer_scene.h"
+#include "editor/import/3d/resource_importer_scene.h"
 
 class Animation;
 class Node;
